move the helper function
