+++ Framework/PythonInterface/plugins/algorithms/component_info_utils.py
